requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
neo4j==5.25.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.0
httpx==0.27.0
openai>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
